Added with_each to global functions Added FlexTable interface
