+++ templates/debug/inputs.html
@@ -194,6 +194,7 @@
                     <option value="swift">Swift</option>
                     <option value="vala">Vala</option>
                 </input-autocomplete>
+
             </div>
             <h3>&lt;input-array/&gt; <help-span
                     value="This component allows you to build, restore, and submit an array.">
